Introduce the "primary_master" group.
This change adds a primary_master host group so that we are able to
rebuild from any combination of nodes. We delegate a node in from the
'masters' group the 'primary_master' and use it to bootstrap the others.

Update provision.py to randomly select a primary master from the group
of masters.

Signed-off-by: Craig Tracey <craigtracey@gmail.com>