Move packages from the common tasks to defaults
Instead of hardcoding the packages in the common tasks themselves, move
them to the defaults for the common role. This allows them to be
modified selectively during configuration.

Introduce the "extra" packages concept where users may include
additional packages over top of the reauired set.

Fix the install of EPEL to install the rpm from over the wire.

Signed-off-by: Craig Tracey <craigtracey@gmail.com>