Fix etcd install failure when re-run
The `creates` parameter on unarchive module causes it to be skipped when
re-run after a successful install.  However, the next copy step tries to run
again even though the downloads are not present.  This change will not
run the step to "move binaries into path" if the previous step is
skipped.

Signed-off-by: Richard Lander <lander2k2@protonmail.com>