Introduce the modify_manifest ansible module
This module allows for the http(s) get of a manifest (ie. those supplied
by Calico and/or Canal), and allows the developer to apply inline
changes to the file with jsonpath syntax. This is then output to the
path of the developer's choosing. If conditionals and/or modifications
do not match, the file will simply be streamed to a file, unchanged.

Signed-off-by: Craig Tracey <craigtracey@gmail.com>