Completely refactor the wardroom cli tool
It had previously been that the wardroom tool had been buried in the
swizzle directory and that it had only been responsible for aiding in
provisioning. Now it is also capable of helping to build the images as
well.

Further this makes the whole repo into a full encapsulated python
module as well as Docker container.

Signed-off-by: Craig Tracey <craigtracey@gmail.com>