Make playbooks more composable
Break out the playbooks such that users can selectively run parts of the
code. In particular, the act of adding nodes should be pretty
lightweight. Whereas currently we run the whole playbook to add nodes,
this introduces an add-nodes.yml playbook so that users do not need to
touch the control plane nodes.

Signed-off-by: Craig Tracey <craigtracey@gmail.com>